backend/requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic
pydantic-settings
python-dotenv
twilio
sendgrid
email-validator
pytz
